Fixed issue when calling globalOutputStream() with 64 bit data types like std::size_t.

+++ plugins/mapdoom3/NodeExporter.cpp
@@ -50,7 +50,7 @@ NodeExporter::~NodeExporter() {
 	// Write the closing braces of the information file
 	_infoStream << "}\n";
 
-	globalOutputStream() << _layerInfoCount << " node-to-layer mappings written.\n";
+	globalOutputStream() << static_cast<int>(_layerInfoCount) << " node-to-layer mappings written.\n";
 }
 
 // Pre-descent callback
